backend/requirements.txt
fastapi
uvicorn
python-multipart
spacy
pdfplumber
python-docx
pydantic
